refactor: switch components to kub-case
