+++ src/components/command-table-row.tsx
