+++ src/components/ping-table.tsx
